add part 2
Part 2 is based off of part 1. While loop has been gutted. Plans are to make a line buffer, pass it to a function.
This function will make another buffer which will hold all characters from buffer1 to i, and it will be checked against
a switch statement of number words. The first loop will check for the first number and immediately store it.
After the first number is stored, the program will store all subsequent numbers in the second variable. After a number is confirmed
i will be decremented by 1 to make sure a scenario such as "...twone" does not occur. This makes the last checked char
get checked again when starting the search for the next number.